Make discover_tools state its trigger points and withhold setup until select
Measured baselines (docs/DISCOVERY_RATE_BENCHMARK.md) showed browse recall of
0-38% and a select rate of 0% across every model tested, with capable models
wiring up vendor CLIs and SDKs in up to 45% of trials without ever calling
Discovery. The trigger policy lives only in this tool's schema, since
prompt_tests asserts Discovery is never injected into the system prompt, so
that is the only lever.

The description now names the concrete moments to browse (before installing a
vendor SDK or CLI, before vendor API calls or config, before fetching vendor
docs or pricing, before connecting an MCP server, before recommending a
provider), states the select obligation, and draws negative scope so purely
local work does not trigger it.

The browse listing no longer prints each entry's setup instructions. That was
the direct cause of the 0% select rate: the agent already had everything it
needed, so the second half of browse-then-select never happened. Setup now
lives only in the select response, and the listing leads with that next step.

Tests pin the contract rather than the phrasing. The secret scanner moves to
discover_secrets.rs, keeping discover.rs under its size ratchet.
